requirements.txt
Django==4.2
PySocks==1.7.1
requests==2.25.1
django-environ==0.9.0
gunicorn==20.1.0
gunicorn[eventlet]>=0.33
gunicorn[gevent]>=1.4
gunicorn[tornado]>=0.2
djangorestframework==3.15.2